// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck
/*
 * This file exports all enum related types from the schema.
 *
 * 🟢 You can import this file directly.
 */

export const MessageRole = {
  user: "user",
  assistant: "assistant",
  system: "system",
} as const;

export type MessageRole = (typeof MessageRole)[keyof typeof MessageRole];